requirements.txt
streamlit
python-pptx
langchain
langchain-community
langchain-openai
python-dotenv